feat: move app created logic into finalize step on new app page
