+++ frontend/src/components/connections/SuggestedAppData.tsx
@@ -62,7 +62,9 @@ export type AppStoreApp = {
   chromeLink?: string;
   firefoxLink?: string;
 
-  guide?: React.ReactNode;
+  installGuide?: React.ReactNode;
+  finalizeGuide?: React.ReactNode;
+  hideConnectionQr?: boolean;
   internal?: boolean;
 };
 
@@ -199,57 +201,36 @@ export const appStoreApps: AppStoreApp[] = (
       logo: alby,
       extendedDescription:
         "Connect your Hub to lightning-enabled websites and lets you pay seamlessly on the web",
-      guide: (
+      hideConnectionQr: true,
+      installGuide: (
         <>
           <div>
-            <h3 className="font-medium">In Alby Browser Extension</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
-                1. Download and open{" "}
+                Download and open{" "}
                 <span className="font-medium text-foreground">
                   Alby Extension
                 </span>{" "}
                 in your desktop browser or for Firefox mobile
               </li>
-              <li>2. Set your Unlock Passcode</li>
+              <li>Set your Unlock Passcode</li>
               <li>
-                3. Connect your wallet via the{" "}
+                Choose{" "}
                 <span className="font-medium text-foreground">
-                  Alby Account
-                </span>{" "}
-                (recommended) or via{" "}
-                <span className="font-medium text-foreground">
-                  Nostr Wallet Connect
-                </span>{" "}
-                (Find Your Wallet → Nostr Wallet Connect)
+                  Bring your own Wallet {"->"} Alby Hub
+                </span>
               </li>
             </ul>
           </div>
+        </>
+      ),
+      finalizeGuide: (
+        <>
           <div>
-            <h3 className="font-medium">
-              In Alby Hub{" "}
-              <span className="text-muted-foreground font-normal">
-                (only needed if you connect via NWC)
-              </span>
-            </h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
+              <li>Copy the connection secret below</li>
               <li>
-                4. Click{" "}
-                <Link
-                  to="/apps/new?app=alby-extension"
-                  className="font-medium text-foreground underline"
-                >
-                  Connect to Alby Extension
-                </Link>
-              </li>
-              <li>5. Set app's wallet permissions (full access recommended)</li>
-            </ul>
-          </div>
-          <div>
-            <h3 className="font-medium">In Alby Browser Extension</h3>
-            <ul className="list-inside text-muted-foreground">
-              <li>
-                6. Paste the connection secret from Alby Hub →{" "}
+                Paste into the extension and then click{" "}
                 <span className="font-medium text-foreground">Continue</span>
               </li>
             </ul>
@@ -267,44 +248,28 @@ export const appStoreApps: AppStoreApp[] = (
       logo: damus,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      installGuide: (
+        <>
+          <p className="text-muted-foreground">
+            Download and open{" "}
+            <span className="font-medium text-foreground">Damus</span> on your
+            iOS device
+          </p>
+        </>
+      ),
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Damus</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
-                1. Download and open{" "}
-                <span className="font-medium text-foreground">Damus</span> on
-                your iOS device
-              </li>
-              <li>
-                2. Go to{" "}
+                Go to{" "}
                 <span className="font-medium text-foreground">Wallet</span> →{" "}
                 <span className="font-medium text-foreground">
                   Attach Wallet
                 </span>
               </li>
-            </ul>
-          </div>
-          <div>
-            <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
-              <li>
-                3. Click{" "}
-                <Link
-                  to="/apps/new?app=damus"
-                  className="font-medium text-foreground underline"
-                >
-                  Connect to Damus
-                </Link>
-              </li>
-              <li>4. Set app's wallet permissions (full access recommended)</li>
-            </ul>
-          </div>
-          <div>
-            <h3 className="font-medium">In Damus</h3>
-            <ul className="list-inside text-muted-foreground">
-              <li>5. Scan or paste the connection secret from Alby Hub</li>
+              <li>Scan or paste the connection secret from Alby Hub</li>
             </ul>
           </div>
         </>
@@ -321,11 +286,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: amethyst,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Amethyst</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download and open{" "}
                 <span className="font-medium text-foreground">Amethyst</span> on
@@ -347,7 +312,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -362,7 +327,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Amethyst</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Scan or paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -381,11 +346,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: primal,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Primal</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download and open{" "}
                 <span className="font-medium text-foreground">Primal</span> on
@@ -407,7 +372,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -425,7 +390,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Primal</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 5. Click on "Paste NWC String" or "Scan NWC QR Code" to connect
                 Alby Hub
@@ -444,11 +409,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: zapstream,
       extendedDescription:
         "Tip streamers, zap comments and pay or receive sats for streaming time with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Zap Stream</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -472,7 +437,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -487,7 +452,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Zap Stream</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 5. Paste connection secret from Alby Hub and click on{" "}
                 <span className="font-medium text-foreground">Connect</span>
@@ -506,7 +471,7 @@ export const appStoreApps: AppStoreApp[] = (
       logo: btcpay,
       extendedDescription:
         "Receive payments directly to your Hub for products you sell online",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <p>
@@ -526,7 +491,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In BTCPay Server</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>1. Log in to your BTCPay Server dashboard</li>
               <li>
                 2. Find connection configuration for your Lightning node (
@@ -547,7 +512,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -566,7 +531,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In BTCPay Server</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 7. Paste the connection secret (nostr+walletconnect://....) in
                 the configuration field
@@ -589,7 +554,7 @@ export const appStoreApps: AppStoreApp[] = (
       logo: lnbits,
       extendedDescription:
         "Connect your Alby Hub to LNbits to give extra functionality through plugins such as BOLT cards and lightning vouchers",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <p>
@@ -606,7 +571,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In LNbits</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>1. Log in to your LNbits admin dashboard</li>
               <li>
                 2. Go to{" "}
@@ -629,7 +594,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -645,7 +610,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In LNbits</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 7. Paste the connection secret (nostr+walletconnect://....)
                 under{" "}
@@ -672,7 +637,7 @@ export const appStoreApps: AppStoreApp[] = (
       logo: coracle,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <p>
             You can connect your Alby Hub to Coracle to zap Nostr notes directly
@@ -680,7 +645,7 @@ export const appStoreApps: AppStoreApp[] = (
           </p>
           <div>
             <h3 className="font-medium">In Coracle</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Log in with your Nostr keys to{" "}
                 <ExternalLink
@@ -710,7 +675,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -740,7 +705,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Coracle</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 7. Paste the connection secret (nostr+walletconnect://....) and
                 click{" "}
@@ -760,7 +725,7 @@ export const appStoreApps: AppStoreApp[] = (
       logo: nostter,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <p>
             You can connect your Alby Hub to Nostter to zap Nostr notes directly
@@ -768,7 +733,7 @@ export const appStoreApps: AppStoreApp[] = (
           </p>
           <div>
             <h3 className="font-medium">In Nostter</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Log in with your Nostr keys to{" "}
                 <ExternalLink
@@ -787,7 +752,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -817,7 +782,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Nostter</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 6. Paste the connection secret (nostr+walletconnect://....)
                 under{" "}
@@ -851,11 +816,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: wavlake,
       extendedDescription:
         "Support artists by paying to upvote music you enjoy with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Wavlake</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download and open{" "}
                 <span className="font-medium text-foreground">Wavlake</span> on
@@ -873,7 +838,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -888,7 +853,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Wavlake</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Scan or paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -906,11 +871,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: wavespace,
       extendedDescription:
         "The world's first Bitcoin VISA Debit Card that allows you to spend BTC globally, anywhere VISA is accepted – straight from the safety of your own NWC-enabled wallet. ✨ EXCLUSIVE ALBYHUB SPECIAL🐝 → Get 21% cashback on your wavecard transactions (up to 10,000 sats) using code »AlbyHub«",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In wave.space</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -941,7 +906,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -956,7 +921,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In wave.space</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -972,11 +937,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: snort,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Snort</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1002,7 +967,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1017,7 +982,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Snort</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1033,11 +998,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: hablanews,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Habla News</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1065,7 +1030,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -1080,7 +1045,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Habla News</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1096,11 +1061,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: iris,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Iris</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1124,7 +1089,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -1139,7 +1104,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Iris</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1155,11 +1120,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: clams,
       extendedDescription:
         "Get insights into your transaction history and accounting tools by connecting your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Clams</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download and open{" "}
                 <ExternalLink
@@ -1175,7 +1140,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 {" "}
                 3. Click{" "}
@@ -1191,7 +1156,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Clams</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Add label & paste connection secret</li>
               <li>6. Click Connect and Save</li>
             </ul>
@@ -1208,11 +1173,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: nostrcheckserver,
       extendedDescription:
         "Enable payments to your Hub from users who register or upload and download files",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Click{" "}
                 <Link
@@ -1227,7 +1192,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Nostrcheck server</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Go to{" "}
                 <span className="font-medium text-foreground">Settings</span>{" "}
@@ -1260,11 +1225,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: nostrudel,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Nostrudel</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1293,7 +1258,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1308,7 +1273,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Nostrudel</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1327,11 +1292,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: yakihonne,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In YakiHonne</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1356,7 +1321,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1371,7 +1336,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In YakiHonne</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1397,11 +1362,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: zapplepay,
       extendedDescription:
         "ZapplePay will make payments from your Hub to zap posts when you react to them",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Zapple Pay</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1423,7 +1388,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -1438,7 +1403,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Zapple Pay</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1454,11 +1419,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: lume,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Lume</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download Lume from{" "}
                 <ExternalLink
@@ -1487,7 +1452,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1502,7 +1467,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Lume</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1518,7 +1483,7 @@ export const appStoreApps: AppStoreApp[] = (
       logo: nakapay,
       extendedDescription:
         "Accept Bitcoin Lightning payments directly to your Hub wallet. NakaPay generates invoices that customers pay directly to you - your funds never go through NakaPay (fully non-custodial)",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <p className="text-sm text-muted-foreground mb-4">
@@ -1532,7 +1497,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">Connect Your Alby Hub to NakaPay</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Click{" "}
                 <Link
@@ -1560,7 +1525,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">Set Up Your NakaPay Account</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Visit{" "}
                 <ExternalLink
@@ -1585,7 +1550,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">Complete Non-Custodial Setup</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 7. Paste your Alby Hub NWC connection string in the wallet
                 connection field
@@ -1616,11 +1581,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: zappybird,
       extendedDescription:
         "Makes a payment from your Hub each time your bird flaps its wings",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Zappy Bird</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1648,7 +1613,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1663,7 +1628,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Zappy Bird</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1693,11 +1658,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: nostur,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Nostur</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Download and open{" "}
                 <span className="font-medium text-foreground">Nostur</span> on
@@ -1719,7 +1684,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1734,7 +1699,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Nostur</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1750,11 +1715,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: wherostr,
       extendedDescription:
         "Tip nostr posts and profiles and receive zaps seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Wherostr</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1781,7 +1746,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1796,7 +1761,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Wherostr</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Scan or paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1812,11 +1777,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: stackernews,
       extendedDescription:
         "Upvote posts with sats and receive sats for your own posts directly in your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In stacker news</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1842,7 +1807,7 @@ export const appStoreApps: AppStoreApp[] = (
             <h3 className="font-medium">
               In Alby Hub: Configure the connection secret for sending
             </h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -1857,7 +1822,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In stacker news</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 5. Paste the connection secret from Alby Hub into{" "}
                 <span className="font-medium text-foreground">
@@ -1870,7 +1835,7 @@ export const appStoreApps: AppStoreApp[] = (
             <h3 className="font-medium">
               In Alby Hub: Configure the connection secret for receiving
             </h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 6. Click{" "}
                 <Link
@@ -1893,7 +1858,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In stacker news</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 8. Paste the connection secret from Alby Hub into{" "}
                 <span className="font-medium text-foreground">
@@ -1914,11 +1879,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: paperScissorsHodl,
       extendedDescription:
         "Uses your Hub to pay to play a round, and receive the reward if you win",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Paper Scissors HODL</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -1942,7 +1907,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -1957,7 +1922,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Paper Scissors HODL</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -1973,11 +1938,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: pullthatupjamie,
       extendedDescription:
         "Pay from your Hub to do private AI-powered searches",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Pull That Up Jamie!</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Open{" "}
                 <ExternalLink
@@ -2001,7 +1966,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 4. Click{" "}
                 <Link
@@ -2016,7 +1981,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Pull That Up Jamie!</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>6. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -2032,11 +1997,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: zapstore,
       extendedDescription:
         "Pay to zap apps and support their creators seamlessly with your Hub",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In Zapstore</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Install{" "}
                 <ExternalLink
@@ -2052,7 +2017,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -2067,7 +2032,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Zapstore</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>5. Paste the connection secret from Alby Hub</li>
             </ul>
           </div>
@@ -2086,11 +2051,11 @@ export const appStoreApps: AppStoreApp[] = (
       logo: zeus,
       extendedDescription:
         "Send and receive payments, to and from your Hub, on the go",
-      guide: (
+      finalizeGuide: (
         <>
           <div>
             <h3 className="font-medium">In ZEUS</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 1. Install{" "}
                 <ExternalLink
@@ -2106,7 +2071,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In Alby Hub</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 3. Click{" "}
                 <Link
@@ -2121,7 +2086,7 @@ export const appStoreApps: AppStoreApp[] = (
           </div>
           <div>
             <h3 className="font-medium">In ZEUS</h3>
-            <ul className="list-inside text-muted-foreground">
+            <ul className="list-inside list-decimal text-muted-foreground">
               <li>
                 5. Scan the connection QR using the QR scanner in the bottom
                 right corner of the bottom of the app, OR manually paste in the
